+++ caldavzap_0.13.2/config_rc.js
@@ -1,17 +1,30 @@
+if('serviceWorker' in navigator) {
+	window.addEventListener('load', function() {
+		navigator.serviceWorker.register('./sw.js').then(function(registration) {
+			console.log('ServiceWorker registered');
+		}, function(err) {
+			console.log('ServiceWorker registration failed: ', err);
+		});
+	});
+}
+
 var globalNetworkRoundcubeSettings={
 	href: location.protocol+'//'+location.hostname+(location.port ? ':'+location.port : '')+location.pathname.replace(RegExp('roundcube\.html$'),'')+'auth/',
 	timeOut: 30000
 };
-var globalBackgroundSync=true;
+var globalBackgroundSync=false;
 var globalSyncResourcesInterval=120000;
 var globalEnableRefresh=true;
 var globalEnableKbNavigation=true;
 var globalSettingsType='principal-URL';
+var globalEventStartPastLimit=1;
+var globalEventStartFutureLimit=1;
+var globalTodoPastLimit=1;
 
 var ck = document.cookie.split('; ');
 for(var i = 0; i < ck.length; i++) {
 	var c = ck[i];
-	nameEQ = "cdz=";
+	var nameEQ = "cdz=";
 	while (c.charAt(0) === ' ') {
 		c = c.substring(1,c.length);
     }
@@ -25,8 +38,13 @@ for(var i = 0; i < ck.length; i++) {
 		var globalCalendarEndOfBusiness=carr.cdz_eb;
 		var globalWeekendDays=carr.cdz_wd.split(",");
 		var globalActiveView=carr.cdz_wv;
+		var globalTimeZoneSupport=carr.cdz_ts;
+		var globalRewriteTimezoneComponent=carr.cdz_tr;
+		var globalRemoveUnknownTimezone=carr.cdz_td;
+		var ignoreServerSettings = carr.cdz_is;
 	}
 }
+
 var globalSortAlphabet=' 0123456789'+
 	'AÀÁÂÄÆÃÅĀBCÇĆČDĎEÈÉÊËĒĖĘĚFGĞHIÌÍÎİÏĪĮJKLŁĹĽMNŃÑŇOÒÓÔÖŐŒØÕŌ'+
 	'PQRŔŘSŚŠȘșŞşẞTŤȚțŢţUÙÚÛÜŰŮŪVWXYÝŸZŹŻŽ'+
@@ -43,23 +61,17 @@ var globalSearchTransformAlphabet={
 var globalResourceAlphabetSorting=true;
 var globalNewVersionNotifyUsers=[];
 var globalHideInfoMessageAfter=1800;
-var globalEditorFadeAnimation=666;
-var globalEventStartPastLimit=3;
-var globalEventStartFutureLimit=3;
-var globalTodoPastLimit=1;
+var globalEditorFadeAnimation=333;
 var globalLoadedCalendarCollections=[];
 var globalLoadedTodoCollections=[];
 var globalActiveCalendarCollections=[];
 var globalActiveTodoCollections=[];
-var globalCalendarSelected='/default/';
+var globalCalendarSelected='default/';
 var globalTodoCalendarSelected='default/';
 var globalOpenFormMode='single';
 var globalTodoListFilterSelected=['filterAction', 'filterProgress'];
 var globalDefaultEventDuration=60;
-var globalDisplayHiddenEvents=false;
-var globalTimeZoneSupport=false;
-var globalRewriteTimezoneComponent=true;
-var globalRemoveUnknownTimezone=false;
+var globalDisplayHiddenEvents=false;	
 var globalShowHiddenAlarms=false;
 var globalIgnoreCompletedOrCancelledAlarms=true;
 var globalMozillaSupport=true;